+++ spec/unit/typecast_spec.rb
@@ -546,12 +546,9 @@ describe "Type Casting" do
       @course.ends_at.to_i.should eql(Time.utc(2011, 4, 1, 16, 50, 32).to_i)
     end
 
-    if RUBY_VERSION >= "1.9.1"
-      # In ruby 1.8.7 Time.parse will always return a value. D'OH
-      it 'does not typecast non-time values' do
-        @course.ends_at = 'not-time'
-        @course['ends_at'].should eql('not-time')
-      end
+    it 'does not typecast non-time values' do
+      @course.ends_at = 'not-time'
+      @course['ends_at'].should eql('not-time')
     end
   end