+++ history.md
@@ -21,6 +21,7 @@
   * Fixing `belongs_to` setting foreign key cache issue.
   * Support typecasting `Symbol`
   * Added `:array` option to properties
+  * Typecasting Dates, Times, and Booleans, with invalid values returns nil
   
   * API Breaking Changes
     * Properties with blocks are now singular unless the `array: true` option is passed.